exl_load_app/exl_load/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
Django==3.0.7
django-picklefield==3.0.1
loader==2017.9.11
numpy==1.19.0
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
pyxlsb==1.0.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
uWSGI==2.0.19.1
xlrd==1.2.0
